MainView Updates
updated getMovies, was importing all info as string, needed to change to Date for 'birthdays' propType

commented out Genres and Directors in NavBar, not ready for that rendering?

needed to render <MovieCard /> surrounded by <Row> for grid format, some SCSS styling to center it.

updated if statements for paths /directors and /genres so it would reload them, saying (!movies) was never going to not render that line as state movies is an empty array.